+++ Adj/Program.hs
@@ -3,3 +3,4 @@ module Adj.Program (module Exports) where
 import Adj.Program.Instruction as Exports
 import Adj.Program.Construction as Exports
 import Adj.Program.Option as Exports
+import Adj.Program.Progress as Exports