+++ CHANGELOG.md
@@ -79,3 +79,4 @@
 * Define `:+:.` type alias
 * Define `.:*:` type alias
 * Define `.:+:` type alias
+* Define `Adj.Program.Progress` module